Remove weird sentencetree module for method on Sentence.
