+++ pyconll/unit/sentence.py
@@ -6,6 +6,7 @@ import operator
 import re
 
 from pyconll.conllable import Conllable
+from pyconll.tree._treebuilder import TreeBuilder
 from pyconll.unit import Token
 
 
@@ -201,6 +202,64 @@ class Sentence(Conllable):
             raise ValueError('Key cannot be {}'.format(Sentence.TEXT_KEY))
 
         self._meta[key] = value
+
+    def to_tree(self):
+        """
+        Creates a Tree data structure from the current sentence.
+
+        An empty sentence will create a Tree with no data and no children.
+
+        Returns:
+            A constructed Tree that represents the dependency graph of the
+            sentence.
+        """
+        children_tokens = {}
+
+        root_token = None
+        for token in self:
+            parent_key = token.head
+
+            try:
+                children_tokens[parent_key].append(token)
+            except KeyError:
+                children_tokens[parent_key] = [token]
+
+            if token.head == '0':
+                root_token = token
+
+        builder = TreeBuilder()
+        builder.create_root(root_token)
+        if root_token:
+            Sentence._create_tree_helper(builder, self, root_token,
+                                         children_tokens)
+        root = builder.build()
+
+        return root
+
+    @staticmethod
+    def _create_tree_helper(builder, sentence, root, children_tokens):
+        """
+        Method to create a tree from a sentence given the root token.
+
+        Args:
+            builder: The TreeBuilder currently being used to create the Tree.
+            sentence: The sentence to construct the tree from.
+            root: The current token we are constructing the tree at.
+            children_tokens: A dictionary from token id to children tokens.
+
+        Returns:
+            A Tree constructed given the sentence structure.
+        """
+        try:
+            tokens = children_tokens[root.id]
+        except KeyError:
+            tokens = []
+
+        for token in tokens:
+            builder.add_child(data=token, move=True)
+            Sentence._create_tree_helper(builder, sentence, token,
+                                         children_tokens)
+            builder.move_to_parent()
 
     def conll(self):
         """
